+++ src/main/java/org/atdev/artrip/domain/Enum/DomesticRegion.java
@@ -17,6 +17,4 @@ public enum DomesticRegion {
 
     private final String region;
     private final String imageUrl;
-
-
 }
