+++ src/main/java/org/atdev/artrip/domain/keyword/service/KeywordService.java
@@ -1,12 +1,19 @@
 package org.atdev.artrip.domain.keyword.service;
 
 import lombok.RequiredArgsConstructor;
+import org.atdev.artrip.domain.auth.data.User;
+import org.atdev.artrip.domain.auth.repository.UserRepository;
+import org.atdev.artrip.domain.keyword.data.Keyword;
+import org.atdev.artrip.domain.keyword.data.UserKeyword;
 import org.atdev.artrip.domain.keyword.repository.KeywordRepository;
 import org.atdev.artrip.domain.keyword.repository.UserKeywordRepository;
 import org.atdev.artrip.domain.keyword.web.dto.KeywordResponse;
+import org.atdev.artrip.global.apipayload.code.status.UserError;
+import org.atdev.artrip.global.apipayload.exception.GeneralException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -16,6 +23,28 @@ public class KeywordService {
 
     private final KeywordRepository keywordRepository;
     private final UserKeywordRepository userKeywordRepository;
+    private final UserRepository userRepository;
+
+    @Transactional
+    public void saveUserKeywords(Long userId, List<Long> keywordIds) {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new GeneralException(UserError._USER_NOT_FOUND));
+
+        userKeywordRepository.deleteByUser(user);
+
+        List<Keyword> keywords = keywordRepository.findAllById(keywordIds);
+
+        List<UserKeyword> userKeywords = keywords.stream()
+                .map(keyword -> UserKeyword.builder()
+                        .user(user)
+                        .keyword(keyword)
+                        .createdAt(LocalDateTime.now())
+                        .build())
+                .toList();
+
+        userKeywordRepository.saveAll(userKeywords);
+    }
 
     @Transactional
     public List<KeywordResponse> getAllKeywords() {
