+++ src/main/java/org/atdev/artrip/domain/user/web/controller/UserController.java
@@ -2,6 +2,7 @@ package org.atdev.artrip.domain.user.web.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
+import org.atdev.artrip.domain.keyword.service.KeywordService;
 import org.atdev.artrip.domain.keyword.web.dto.KeywordRequest;
 import org.atdev.artrip.domain.user.service.UserService;
 import org.atdev.artrip.domain.user.web.dto.request.NicknameRequest;
@@ -25,22 +26,6 @@ import org.springframework.web.multipart.MultipartFile;
 public class UserController {
 
     private final UserService userService;
-
-    @Operation(summary = "나의 취향 분석", description = "내가 선택한 키워드 선택 저장")
-    @ApiErrorResponses(
-            common = {CommonError._INTERNAL_SERVER_ERROR, CommonError._UNAUTHORIZED},
-            keyword = {KeywordError._KEYWORD_INVALID_REQUEST, KeywordError._KEYWORD_SELECTION_LIMIT_EXCEEDED, KeywordError._KEYWORD_NOT_FOUND}
-    )
-    @PostMapping("/keywords")
-    public ResponseEntity<CommonResponse<Void>> saveUserKeywords(
-            @AuthenticationPrincipal UserDetails userDetails,
-            @RequestBody KeywordRequest request) {
-
-        Long userId = Long.parseLong(userDetails.getUsername()); // subject → userId형변환
-
-        userService.saveUserKeywords(userId, request.getKeywordIds());
-        return ResponseEntity.ok(CommonResponse.onSuccess(null));
-    }
 
     @Operation(summary = "프로필 이미지 추가", description = "프로필 이미지를 추가합니다")
     @PostMapping(value = "/profile", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
